poniendo alertas mas bonitas

+++ zombie-plash/js/validar_inicio_sesion.js
@@ -18,16 +18,24 @@ document.getElementById('formulario').addEventListener('submit', function(e) {
         console.log('Respuesta completa del servidor:', data);
         if (data.success) {
             console.log('Inicio de sesión exitoso:', data.session_info);
-            alert('Inicio de sesión exitoso. Redirigiendo...');
-            localStorage.setItem('id_jugador', data.session_info["id_usuario"]);
-            window.location.href = '../html/inicio.php'; // Asegúrate de que esta ruta sea correcta
+            Swal.fire({
+                title: 'Éxito',
+                text: 'Inicio de sesión exitoso. Redirigiendo...',
+                icon: 'success',
+                confirmButtonText: 'Aceptar'
+            }).then(() => {
+                localStorage.setItem('id_jugador', data.session_info["id_usuario"]);
+                window.location.href = '../html/inicio.php'; // Asegúrate de que esta ruta sea correcta
+            });
         } else {
             // Mostrar error general
-            if (data.errors && data.errors.general) {
-                document.getElementById('generalError').textContent = data.errors.general;
-            } else {
-                document.getElementById('generalError').textContent = 'Ocurrió un error desconocido';
-            }
+            const errorMessage = data.errors && data.errors.general ? data.errors.general : 'Ocurrió un error desconocido';
+            Swal.fire({
+                title: 'Error',
+                text: errorMessage,
+                icon: 'error',
+                confirmButtonText: 'Aceptar'
+            });
         }
         if (data.debug) {
             console.log('Información de depuración:', data.debug);
@@ -35,6 +43,11 @@ document.getElementById('formulario').addEventListener('submit', function(e) {
     })
     .catch(error => {
         console.error('Error:', error);
-        document.getElementById('generalError').textContent = 'Error de conexión con el servidor';
+        Swal.fire({
+            title: 'Error',
+            text: 'Error de conexión con el servidor',
+            icon: 'error',
+            confirmButtonText: 'Aceptar'
+        });
     });
 });
